Added manhattanGrade for HillClimbing

+++ src/main/java/algorithms/Searchable.java
@@ -1,14 +1,15 @@
 package algorithms;
 
-import models.PuzzleState;
 import models.State;
 
 import java.util.ArrayList;
 
-public interface Searchable <T> {
+public interface Searchable<T> {
     State<T> getInitialState();
+
     boolean isGoal(State<T> state);
+
     ArrayList<State<T>> getAllPossibleStates(State<T> s);
 
-    double grade(State state);
+    int grade(State state);
 }
